Extract the Session mailbox protocol, durable rails and pump
Issue #140 moves the complete Session mailbox vertical out of raw networking as
one coherent slice, absorbing the bounded relocations #191 and #190. The
1,545-line `wow_network::player_registry` and the pump that drained it become
private `wow_world::session::mailbox` submodules: `protocol` (the
`SessionCommand` enum and every ordinary payload, plus the shared
`Player::m_clientGUIDs` handle), `durable` (the creature-runtime rail and the
loot-money admission/fence/guard), and `pump` (the single consumer).

Queue identity, FIFO order, durability, bounded capacity, incarnation fences,
acknowledgements and shutdown drain are unchanged. The pump keeps its exact
draining order -- durable creature rail up to its first visibility-gated packet,
then the bounded general rail, then the deferred durable suffix -- and still
disconnects a session whose authoritative backlog overflowed rather than
dropping transitions. All 13 mailbox tests move with the owner and pass.

Unlike the two moves before it, this is a net dependency reduction. With the
mailbox gone, `wow-network` no longer references `wow-data`, `wow-loot` or
`wow-social` at all: the only remaining mentions were two string literals in
the listener's own boundary assertion. All three edges are deleted and their
issue-linked exceptions retired, including the transitional
`wow-network -> wow-social` edge #137 had declared and assigned here. The
workspace falls from 109 to 106 edges and from 20 to 17 baseline exceptions,
and raw networking now depends only on transport crates. `check_architecture`'s
own self-test had pinned `wow-network -> wow-data` as its example of a debt
edge; it is repointed at a live one.

The one real surface change is that 29 command handlers in `handlers/loot.rs`
become `pub(crate)`. The pump that dispatches to them now lives in a sibling
module and Rust has no narrower visibility that reaches it; they remain
crate-internal and unexported. The `WorldSession` surface is otherwise
untouched: 738 fields and 3,341 associated items, with no method added or
removed, one field type path following the move.

Guards follow the owner. `session-ownership-check` now scans
`crate::session::mailbox`, without which `SessionCommand` and its 41 payload
types would have left the pinned inventory entirely. The anti-glob registry
guard gains `mailbox`, with a new adversarial test covering both the crate-root
and submodule forms. Reviewed baseline delta: 37 variants and 42 payload types
keep identical names and shapes with only their modules moving, 44 generated
inputs remap the same way, and the 71 bridge identities and 570 direct-registry
rows are unchanged.

Validation: cargo check --workspace --all-targets clean with no new warnings;
cargo test -p wow-world --lib mailbox 13/13; -p wow-network --lib 32/32
(exactly the 45 minus the 13 that moved); -p wow-social --lib 80/80;
-p world-server --lib 449/449; handler-contract-check registry_inventory
rejection suite 4/4; session-ownership-check check --syntax-only;
check_architecture.py check and self-test; cargo fmt --all -- --check and
git diff --check.

Closes #140
