+++ crates/world-server/src/lib.rs
@@ -48,12 +48,7 @@ use wow_loot::{
 };
 use wow_network::session_mgr::SessionManager;
 use wow_network::world_socket::{AccountInfo, AccountLookup};
-use wow_network::{
-    GameEventQuestCompleteCommandLikeCpp, GameEventQuestCompleteResponseLikeCpp,
-    KickLikeCppCommand, ResetSeasonalQuestStatusCommand, SendVisibleObjectValuesUpdateCommand,
-    SessionCommand, SocketTimeoutsLikeCpp, WorldListenerPolicyLikeCpp,
-    WorldSessionShutdownFlushLikeCppCommand, WorldSessionShutdownFlushResultLikeCpp,
-};
+use wow_network::{SocketTimeoutsLikeCpp, WorldListenerPolicyLikeCpp};
 use wow_packet::{
     ServerPacket,
     packets::chat::{ChatMsg, ChatPkt},
@@ -64,6 +59,12 @@ use wow_social::group::{
     load_groups_from_db_rows_like_cpp, tick_all_group_ready_checks_like_cpp,
 };
 use wow_world::session::directory::PlayerRegistry;
+use wow_world::session::mailbox::{
+    GameEventQuestCompleteCommandLikeCpp, GameEventQuestCompleteResponseLikeCpp,
+    KickLikeCppCommand, ResetSeasonalQuestStatusCommand, SendVisibleObjectValuesUpdateCommand,
+    SessionCommand, WorldSessionShutdownFlushLikeCppCommand,
+    WorldSessionShutdownFlushResultLikeCpp,
+};
 use wow_world::{
     BattlePetAccountRegistryLikeCpp, ChatFloodConfigLikeCpp, ChatLevelRequirementsLikeCpp,
     ChatListenRangesLikeCpp, LoginBattlePetPersistenceLikeCpp, LootDropRatesLikeCpp,
