+++ crates/wow-world/src/session/mailbox/protocol.rs
@@ -3,26 +3,24 @@
 // Based on TrinityCore protocol research (https://github.com/TrinityCore/TrinityCore)
 // Licensed under GPL v3 — https://www.gnu.org/licenses/gpl-3.0.html
 
-//! Session mailbox protocol and durable rails for active player sessions.
+//! Ordinary Session mailbox protocol: the cross-session command enum and every
+//! payload a producer may enqueue on the owning session.
 //!
-//! The commands here are the cross-session mutation and delivery protocol a
-//! producer enqueues on the owning session, plus the durable creature-runtime
-//! and loot-money rails that must survive backpressure. The connected-session
-//! directory that resolves these endpoints is owned by
-//! `wow_world::session::directory`; the pump and the mailbox boundary itself
-//! move out under issue #140.
+//! Enqueueing is the only way another task mutates this session; the payloads
+//! carry committed facts, never borrowed state or transport handles.
 
-use std::collections::{HashSet, VecDeque};
-use std::sync::{
-    Arc, Mutex,
-    atomic::{AtomicBool, AtomicU64},
-};
+use std::collections::HashSet;
+use std::sync::Arc;
+use std::sync::atomic::{AtomicBool, AtomicU64};
 use std::time::Instant;
+
 use wow_core::ObjectGuid;
 use wow_loot::{LootClaimLease, OwnedLootAuthority};
 use wow_packet::packets::loot::LootEntry;
 use wow_packet::packets::party::PartyUpdate;
 pub use wow_social::group::GroupDifficultyKindLikeCpp;
+
+use super::durable::DurableLootMoneyPersistenceTrackerLikeCpp;
 
 /// C++ `Player::m_clientGUIDs` held behind a shared handle.
 ///
@@ -382,93 +380,6 @@ pub struct ReconcilePvpCombatExpiryLikeCppCommand {
     pub instance_id: u32,
 }
 
-/// Durable FIFO handoff for map-owned creature transitions that have
-/// already committed authoritative state.
-///
-/// The bounded general-purpose session queue may legitimately reject visual
-/// fanout under backpressure. These commands cannot be dropped, but the global
-/// map tick also cannot wait for a stalled session. C++ publishes every melee
-/// swing and attack transition in order, so this rail retains each committed
-/// event until the owning session drains it. A session that cannot drain a
-/// bounded backlog is marked desynchronized and disconnected instead of
-/// silently losing authoritative events or growing memory without limit.
-pub const MAX_DURABLE_CREATURE_RUNTIME_COMMANDS_LIKE_CPP: usize = 4_096;
-
-#[derive(Default)]
-pub struct DurableCreatureRuntimeCommandsLikeCpp {
-    commands: VecDeque<SessionCommand>,
-    overflowed: bool,
-}
-
-impl DurableCreatureRuntimeCommandsLikeCpp {
-    fn publish_like_cpp(&mut self, command: SessionCommand) -> bool {
-        if self.commands.len() >= MAX_DURABLE_CREATURE_RUNTIME_COMMANDS_LIKE_CPP {
-            self.overflowed = true;
-            return false;
-        }
-        self.commands.push_back(command);
-        true
-    }
-
-    pub fn publish_attack_start_like_cpp(
-        &mut self,
-        command: CreatureAttackStartLikeCppCommand,
-    ) -> bool {
-        self.publish_like_cpp(SessionCommand::CreatureAttackStartLikeCpp(command))
-    }
-
-    pub fn publish_attack_stop_like_cpp(
-        &mut self,
-        command: CreatureAttackStopLikeCppCommand,
-    ) -> bool {
-        self.publish_like_cpp(SessionCommand::CreatureAttackStopLikeCpp(command))
-    }
-
-    pub fn publish_pvp_combat_expiry_like_cpp(
-        &mut self,
-        command: ReconcilePvpCombatExpiryLikeCppCommand,
-    ) -> bool {
-        self.publish_like_cpp(SessionCommand::ReconcilePvpCombatExpiryLikeCpp(command))
-    }
-
-    pub fn publish_melee_damage_like_cpp(
-        &mut self,
-        command: ApplyCreatureMeleeDamageLikeCppCommand,
-    ) -> bool {
-        self.publish_like_cpp(SessionCommand::ApplyCreatureMeleeDamageLikeCpp(command))
-    }
-
-    pub fn publish_send_if_visible_like_cpp(
-        &mut self,
-        command: SendIfVisibleLikeCppCommand,
-    ) -> bool {
-        self.publish_like_cpp(SessionCommand::SendIfVisibleLikeCpp(command))
-    }
-
-    /// Publish START+GO as one queue element so capacity checks and session
-    /// drains cannot observe only one half of a committed spell cast.
-    pub fn publish_creature_spell_cast_if_visible_like_cpp(
-        &mut self,
-        command: SendCreatureSpellCastIfVisibleLikeCppCommand,
-    ) -> bool {
-        self.publish_like_cpp(SessionCommand::SendCreatureSpellCastIfVisibleLikeCpp(
-            command,
-        ))
-    }
-
-    pub fn drain_like_cpp(&mut self) -> Vec<SessionCommand> {
-        self.commands.drain(..).collect()
-    }
-
-    pub fn take_overflowed_and_discard_like_cpp(&mut self) -> bool {
-        let overflowed = std::mem::take(&mut self.overflowed);
-        if overflowed {
-            self.commands.clear();
-        }
-        overflowed
-    }
-}
-
 /// Payload for [`SessionCommand::SendIfVisibleLikeCpp`].
 ///
 /// Carries both `map_id` and `instance_id` so the per-session gate can reject
@@ -816,264 +727,6 @@ pub struct ApplyLootMoneyLikeCppCommand {
     pub published: Arc<AtomicBool>,
 }
 
-/// Durable result of one character-row money mutation.
-///
-/// The detached SQL worker records this before publishing its session command.
-/// Both paths share [`Self::applied`], so logout reconciliation and normal
-/// command delivery cannot apply the same durable delta twice.
-#[derive(Clone, Debug)]
-pub struct DurableLootMoneyCompletionLikeCpp {
-    pub durable_money_before: u64,
-    pub durable_money_after: u64,
-    pub durable_applied_amount: u64,
-    pub applied: Arc<AtomicBool>,
-}
-
-#[derive(Debug, Default)]
-struct DurableLootMoneyPersistenceStateLikeCpp {
-    in_flight: usize,
-    completions: Vec<DurableLootMoneyCompletionLikeCpp>,
-    indeterminate: bool,
-    admission_closed: bool,
-    permanently_closed: bool,
-    active_save_fences: usize,
-}
-
-/// Per-character fence for detached loot-money transactions.
-///
-/// A tracker is published in [`PlayerBroadcastInfo`] and registered directly
-/// by a source session before it opens a transaction for every recipient. This
-/// avoids command acknowledgements (and their A↔B deadlocks), while allowing a
-/// target session to wait for and reconcile durable completions before an
-/// absolute `Player::SaveToDB` money write.
-#[derive(Debug)]
-pub struct DurableLootMoneyPersistenceTrackerLikeCpp {
-    state: Mutex<DurableLootMoneyPersistenceStateLikeCpp>,
-    changed: tokio::sync::watch::Sender<u64>,
-    money_mutation_serial: Arc<tokio::sync::Mutex<()>>,
-}
-
-impl Default for DurableLootMoneyPersistenceTrackerLikeCpp {
-    fn default() -> Self {
-        let (changed, _) = tokio::sync::watch::channel(0);
-        Self {
-            state: Mutex::new(DurableLootMoneyPersistenceStateLikeCpp::default()),
-            changed,
-            money_mutation_serial: Arc::new(tokio::sync::Mutex::new(())),
-        }
-    }
-}
-
-impl DurableLootMoneyPersistenceTrackerLikeCpp {
-    /// Serialize DB money mutations for this character across stored-item and
-    /// group payouts. Multi-recipient workers acquire these locks in sorted
-    /// GUID order before taking the matching character-row locks.
-    pub async fn lock_money_mutation_like_cpp(&self) -> tokio::sync::OwnedMutexGuard<()> {
-        Arc::clone(&self.money_mutation_serial).lock_owned().await
-    }
-
-    #[must_use]
-    pub fn begin_like_cpp(
-        self: &Arc<Self>,
-    ) -> Result<DurableLootMoneyPersistenceGuardLikeCpp, DurableLootMoneyAdmissionClosedLikeCpp>
-    {
-        let mut state = self
-            .state
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        if state.admission_closed {
-            return Err(DurableLootMoneyAdmissionClosedLikeCpp);
-        }
-        state.in_flight += 1;
-        drop(state);
-        let _ = self.changed.send_modify(|generation| {
-            *generation = generation.wrapping_add(1);
-        });
-        Ok(DurableLootMoneyPersistenceGuardLikeCpp {
-            tracker: Arc::clone(self),
-            resolved: false,
-        })
-    }
-
-    /// Close admission before observing `in_flight` and keep it closed across
-    /// snapshot plus SQL commit. Either a source registers first and the save
-    /// waits, or the save closes first and the source fails before BEGIN.
-    #[must_use]
-    pub fn close_admission_for_save_like_cpp(self: &Arc<Self>) -> DurableLootMoneySaveFenceLikeCpp {
-        let mut state = self
-            .state
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        state.active_save_fences = state.active_save_fences.saturating_add(1);
-        state.admission_closed = true;
-        DurableLootMoneySaveFenceLikeCpp {
-            tracker: Arc::clone(self),
-        }
-    }
-
-    /// Logout closes the old registry-published tracker permanently. A source
-    /// that cloned it before unregister cannot mutate the character after its
-    /// final save.
-    pub fn close_admission_permanently_like_cpp(&self) {
-        let mut state = self
-            .state
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        state.admission_closed = true;
-        state.permanently_closed = true;
-    }
-
-    pub async fn wait_until_idle_like_cpp(&self) {
-        let mut changed = self.changed.subscribe();
-        loop {
-            if self
-                .state
-                .lock()
-                .unwrap_or_else(|poisoned| poisoned.into_inner())
-                .in_flight
-                == 0
-            {
-                return;
-            }
-            if changed.changed().await.is_err() {
-                return;
-            }
-        }
-    }
-
-    /// Returns every completion whose shared exact-once gate is still open.
-    /// Applied entries are pruned only after their CAS is observable.
-    #[must_use]
-    pub fn pending_completions_like_cpp(&self) -> Vec<DurableLootMoneyCompletionLikeCpp> {
-        let mut state = self
-            .state
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        state.completions.retain(|completion| {
-            !completion
-                .applied
-                .load(std::sync::atomic::Ordering::Acquire)
-        });
-        state.completions.clone()
-    }
-
-    #[must_use]
-    pub fn is_indeterminate_like_cpp(&self) -> bool {
-        self.state
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner())
-            .indeterminate
-    }
-
-    pub fn mark_indeterminate_like_cpp(&self) {
-        {
-            let mut state = self
-                .state
-                .lock()
-                .unwrap_or_else(|poisoned| poisoned.into_inner());
-            state.indeterminate = true;
-            state.admission_closed = true;
-            state.permanently_closed = true;
-        }
-        let _ = self.changed.send_modify(|generation| {
-            *generation = generation.wrapping_add(1);
-        });
-    }
-
-    fn finish_like_cpp(
-        &self,
-        completion: Option<DurableLootMoneyCompletionLikeCpp>,
-        indeterminate: bool,
-    ) {
-        {
-            let mut state = self
-                .state
-                .lock()
-                .unwrap_or_else(|poisoned| poisoned.into_inner());
-            debug_assert!(state.in_flight != 0);
-            state.in_flight = state.in_flight.saturating_sub(1);
-            if let Some(completion) = completion {
-                state.completions.push(completion);
-            }
-            state.indeterminate |= indeterminate;
-            if indeterminate {
-                state.admission_closed = true;
-                state.permanently_closed = true;
-            }
-        }
-        let _ = self.changed.send_modify(|generation| {
-            *generation = generation.wrapping_add(1);
-        });
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct DurableLootMoneyAdmissionClosedLikeCpp;
-
-impl std::fmt::Display for DurableLootMoneyAdmissionClosedLikeCpp {
-    fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter.write_str("character money persistence admission is closed")
-    }
-}
-
-impl std::error::Error for DurableLootMoneyAdmissionClosedLikeCpp {}
-
-#[derive(Debug)]
-pub struct DurableLootMoneySaveFenceLikeCpp {
-    tracker: Arc<DurableLootMoneyPersistenceTrackerLikeCpp>,
-}
-
-impl Drop for DurableLootMoneySaveFenceLikeCpp {
-    fn drop(&mut self) {
-        let mut state = self
-            .tracker
-            .state
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        debug_assert!(state.active_save_fences != 0);
-        state.active_save_fences = state.active_save_fences.saturating_sub(1);
-        if state.active_save_fences == 0 && !state.permanently_closed {
-            state.admission_closed = false;
-        }
-    }
-}
-
-/// RAII registration for one recipient's durable money mutation.
-#[derive(Debug)]
-pub struct DurableLootMoneyPersistenceGuardLikeCpp {
-    tracker: Arc<DurableLootMoneyPersistenceTrackerLikeCpp>,
-    resolved: bool,
-}
-
-impl DurableLootMoneyPersistenceGuardLikeCpp {
-    pub fn commit_like_cpp(&mut self, completion: DurableLootMoneyCompletionLikeCpp) {
-        if self.resolved {
-            return;
-        }
-        self.resolved = true;
-        self.tracker.finish_like_cpp(Some(completion), false);
-    }
-
-    /// A COMMIT was attempted but its outcome could not be reconciled. The
-    /// target must skip absolute money saves until it disconnects/reloads.
-    pub fn mark_indeterminate_like_cpp(&mut self) {
-        if self.resolved {
-            return;
-        }
-        self.resolved = true;
-        self.tracker.finish_like_cpp(None, true);
-    }
-}
-
-impl Drop for DurableLootMoneyPersistenceGuardLikeCpp {
-    fn drop(&mut self) {
-        if !self.resolved {
-            self.tracker.finish_like_cpp(None, false);
-            self.resolved = true;
-        }
-    }
-}
-
 /// Durable `Loot::NotifyMoneyRemoved` delivery for an active viewer that is
 /// not itself a payout recipient.
 #[derive(Clone, Debug)]
@@ -1150,396 +803,4 @@ pub struct LootRollVoteCommand {
     /// Immutable enqueue-time identity of the exact roll instance. C++ queues
     /// no cross-session surrogate; its player retains the exact `LootRoll*`.
     pub roll_identity: LootRollCommandIdentityLikeCpp,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// A producer selecting recipients must never observe a visibility
-    /// transition half applied, nor the old membership while the packets that
-    /// replace it are already queued.
-    #[test]
-    fn visibility_transition_publishes_membership_and_packets_as_one_step_like_cpp() {
-        use std::sync::mpsc;
-        use std::time::Duration;
-
-        let visibility = SharedClientVisibleGuidsLikeCpp::default();
-        let leaving = ObjectGuid::create_player(1, 1);
-        let arriving = ObjectGuid::create_player(1, 2);
-        visibility.insert(leaving);
-
-        let reader_handle = visibility.clone();
-        let (entered_tx, entered_rx) = mpsc::channel();
-        let (observed_tx, observed_rx) = mpsc::channel();
-        let reader = std::thread::spawn(move || {
-            entered_rx
-                .recv()
-                .expect("transition entered its publish step");
-            // This read can only complete once the transition released its write.
-            observed_tx
-                .send(reader_handle.snapshot_like_cpp())
-                .expect("reader reported its snapshot");
-        });
-
-        visibility.publish_transition_like_cpp(
-            |guid| *guid != leaving,
-            [arriving],
-            || {
-                entered_tx.send(()).expect("reader was signalled");
-                std::thread::sleep(Duration::from_millis(50));
-                assert!(
-                    observed_rx.try_recv().is_err(),
-                    "no reader may observe the set while its packets are being published"
-                );
-            },
-        );
-
-        let observed = observed_rx
-            .recv_timeout(Duration::from_secs(5))
-            .expect("reader snapshot");
-        reader.join().expect("reader finished");
-        assert!(
-            !observed.contains(&leaving),
-            "the departed object is gone once the transition is observable"
-        );
-        assert!(
-            observed.contains(&arriving),
-            "the arrived object is present once the transition is observable"
-        );
-    }
-
-    /// Verify that `PlayerBroadcastInfo` carries `instance_id` so that
-
-    /// Verify that `SendIfVisibleLikeCppCommand` carries both `map_id` and
-    /// `instance_id` — required so per-session gate can reject cross-instance
-    /// delivery (Slice 4A.1b).
-    #[test]
-    fn send_if_visible_like_cpp_command_carries_map_and_instance_id() {
-        let guid = ObjectGuid::create_player(1, 7);
-        let cmd = SendIfVisibleLikeCppCommand {
-            queued_at: Instant::now(),
-            source_guid: guid,
-            map_id: 532,
-            instance_id: 99,
-            packet_bytes: vec![0xDE, 0xAD],
-        };
-        assert_eq!(cmd.map_id, 532);
-        assert_eq!(cmd.instance_id, 99);
-    }
-
-    #[test]
-    fn creature_spell_start_and_committed_go_use_one_durable_queue_element_like_cpp() {
-        let source_guid = ObjectGuid::create_world_object(
-            wow_core::guid::HighGuid::Creature,
-            0,
-            1,
-            571,
-            0,
-            123,
-            458,
-        );
-        let command = SendCreatureSpellCastIfVisibleLikeCppCommand {
-            queued_at: Instant::now(),
-            source_guid,
-            map_id: 571,
-            instance_id: 4,
-            start_packet_bytes: vec![0x37, 0x2C, 0xAA],
-            go_packet_bytes: vec![0x36, 0x2C, 0xBB],
-            committed_visibility_like_cpp: SharedClientVisibleGuidsLikeCpp::default(),
-        };
-        let mut durable = DurableCreatureRuntimeCommandsLikeCpp::default();
-
-        assert!(durable.publish_creature_spell_cast_if_visible_like_cpp(command.clone()));
-        let drained = durable.drain_like_cpp();
-
-        assert_eq!(
-            drained.len(),
-            1,
-            "a drain cannot split START from GO because the pair occupies one FIFO element"
-        );
-        let [SessionCommand::SendCreatureSpellCastIfVisibleLikeCpp(drained_command)] =
-            drained.as_slice()
-        else {
-            panic!("expected one atomic spell cast command: {drained:?}");
-        };
-        assert_eq!(drained_command, &command);
-    }
-
-    /// Verify that creature visibility refresh commands are scoped by both map
-    /// and instance. The receiving `WorldSession` applies the same gates before
-    /// forcing its visibility pass.
-    #[test]
-    fn refresh_visible_world_creatures_like_cpp_command_carries_map_and_instance_id() {
-        let cmd = RefreshVisibleWorldCreaturesLikeCppCommand {
-            map_id: 571,
-            instance_id: 7,
-        };
-        assert_eq!(cmd.map_id, 571);
-        assert_eq!(cmd.instance_id, 7);
-    }
-
-    /// Verify the creature melee damage command carries both addressing data
-    /// and final victim health so session delivery can be idempotent.
-    #[test]
-    fn apply_creature_melee_damage_like_cpp_command_carries_final_health() {
-        let attacker = ObjectGuid::create_world_object(
-            wow_core::guid::HighGuid::Creature,
-            0,
-            1,
-            571,
-            0,
-            123,
-            456,
-        );
-        let victim = ObjectGuid::create_player(1, 7);
-        let cmd = ApplyCreatureMeleeDamageLikeCppCommand {
-            attacker_guid: attacker,
-            victim_guid: victim,
-            map_id: 571,
-            instance_id: 3,
-            damage: 11,
-            over_damage: -1,
-            target_level: 80,
-            victim_health_after: 89,
-            victim_health_state_revision_after: 7,
-        };
-
-        assert_eq!(cmd.attacker_guid, attacker);
-        assert_eq!(cmd.victim_guid, victim);
-        assert_eq!(cmd.map_id, 571);
-        assert_eq!(cmd.instance_id, 3);
-        assert_eq!(cmd.victim_health_after, 89);
-        assert_eq!(cmd.victim_health_state_revision_after, 7);
-    }
-
-    #[test]
-    fn creature_attack_start_like_cpp_command_carries_map_and_instance_id() {
-        let attacker = ObjectGuid::create_world_object(
-            wow_core::guid::HighGuid::Creature,
-            0,
-            1,
-            571,
-            0,
-            123,
-            457,
-        );
-        let victim = ObjectGuid::create_player(1, 8);
-        let cmd = CreatureAttackStartLikeCppCommand {
-            attacker_guid: attacker,
-            victim_guid: victim,
-            previous_victim_guid: None,
-            map_id: 571,
-            instance_id: 4,
-            packet_already_broadcast: false,
-        };
-
-        assert_eq!(cmd.attacker_guid, attacker);
-        assert_eq!(cmd.victim_guid, victim);
-        assert_eq!(cmd.map_id, 571);
-        assert_eq!(cmd.instance_id, 4);
-    }
-
-    #[tokio::test]
-    async fn durable_money_save_fence_closes_admission_then_waits_for_prior_worker() {
-        let tracker = Arc::new(DurableLootMoneyPersistenceTrackerLikeCpp::default());
-        let worker = tracker.begin_like_cpp().unwrap();
-        let save_fence = tracker.close_admission_for_save_like_cpp();
-
-        assert!(tracker.begin_like_cpp().is_err());
-        let wait_tracker = Arc::clone(&tracker);
-        let waiter = tokio::spawn(async move {
-            wait_tracker.wait_until_idle_like_cpp().await;
-        });
-        tokio::task::yield_now().await;
-        assert!(!waiter.is_finished());
-
-        drop(worker);
-        tokio::time::timeout(std::time::Duration::from_secs(1), waiter)
-            .await
-            .expect("prior durable worker must release the save wait")
-            .unwrap();
-        assert!(tracker.begin_like_cpp().is_err());
-
-        drop(save_fence);
-        assert!(tracker.begin_like_cpp().is_ok());
-    }
-
-    #[test]
-    fn durable_money_permanent_logout_fence_never_reopens() {
-        let tracker = Arc::new(DurableLootMoneyPersistenceTrackerLikeCpp::default());
-        let save_fence = tracker.close_admission_for_save_like_cpp();
-        tracker.close_admission_permanently_like_cpp();
-        drop(save_fence);
-
-        assert!(tracker.begin_like_cpp().is_err());
-    }
-
-    #[test]
-    fn overlapping_durable_money_save_fences_reopen_only_after_last_drop() {
-        let tracker = Arc::new(DurableLootMoneyPersistenceTrackerLikeCpp::default());
-        let first = tracker.close_admission_for_save_like_cpp();
-        let second = tracker.close_admission_for_save_like_cpp();
-
-        drop(first);
-        assert!(
-            tracker.begin_like_cpp().is_err(),
-            "dropping the first fence must not reopen admission under the second"
-        );
-        drop(second);
-        assert!(tracker.begin_like_cpp().is_ok());
-    }
-
-    #[test]
-    fn durable_money_completion_uses_one_shared_exact_once_gate() {
-        let tracker = Arc::new(DurableLootMoneyPersistenceTrackerLikeCpp::default());
-        let mut worker = tracker.begin_like_cpp().unwrap();
-        let applied = Arc::new(AtomicBool::new(false));
-        worker.commit_like_cpp(DurableLootMoneyCompletionLikeCpp {
-            durable_money_before: 40,
-            durable_money_after: 47,
-            durable_applied_amount: 7,
-            applied: Arc::clone(&applied),
-        });
-
-        let pending = tracker.pending_completions_like_cpp();
-        assert_eq!(pending.len(), 1);
-        assert_eq!(pending[0].durable_money_before, 40);
-        assert_eq!(pending[0].durable_money_after, 47);
-        assert_eq!(pending[0].durable_applied_amount, 7);
-        assert!(
-            pending[0]
-                .applied
-                .compare_exchange(
-                    false,
-                    true,
-                    std::sync::atomic::Ordering::AcqRel,
-                    std::sync::atomic::Ordering::Acquire,
-                )
-                .is_ok()
-        );
-        assert!(tracker.pending_completions_like_cpp().is_empty());
-    }
-
-    #[test]
-    fn indeterminate_money_outcome_permanently_closes_admission_until_relogin() {
-        let tracker = Arc::new(DurableLootMoneyPersistenceTrackerLikeCpp::default());
-        let fence = tracker.close_admission_for_save_like_cpp();
-        tracker.mark_indeterminate_like_cpp();
-        drop(fence);
-
-        assert!(tracker.is_indeterminate_like_cpp());
-        assert!(tracker.begin_like_cpp().is_err());
-    }
-
-    #[test]
-    fn durable_creature_runtime_commands_preserve_committed_fifo_like_cpp() {
-        let victim = ObjectGuid::create_player(1, 900);
-        let attacker = ObjectGuid::create_world_object(
-            wow_core::guid::HighGuid::Creature,
-            0,
-            1,
-            571,
-            0,
-            9001,
-            901,
-        );
-        let mut pending = DurableCreatureRuntimeCommandsLikeCpp::default();
-        assert!(
-            pending.publish_attack_start_like_cpp(CreatureAttackStartLikeCppCommand {
-                attacker_guid: attacker,
-                victim_guid: victim,
-                previous_victim_guid: None,
-                map_id: 571,
-                instance_id: 0,
-                packet_already_broadcast: false,
-            })
-        );
-        assert!(
-            pending.publish_attack_stop_like_cpp(CreatureAttackStopLikeCppCommand {
-                attacker_guid: attacker,
-                victim_guid: victim,
-                map_id: 571,
-                instance_id: 0,
-            })
-        );
-        assert!(pending.publish_pvp_combat_expiry_like_cpp(
-            ReconcilePvpCombatExpiryLikeCppCommand {
-                player_guid: victim,
-                map_id: 571,
-                instance_id: 0,
-            }
-        ));
-        for (victim_health_after, victim_health_state_revision_after) in [(90, 7), (75, 8)] {
-            assert!(pending.publish_melee_damage_like_cpp(
-                ApplyCreatureMeleeDamageLikeCppCommand {
-                    attacker_guid: attacker,
-                    victim_guid: victim,
-                    map_id: 571,
-                    instance_id: 0,
-                    damage: 15,
-                    over_damage: -1,
-                    target_level: 80,
-                    victim_health_after,
-                    victim_health_state_revision_after,
-                }
-            ));
-        }
-
-        let commands = pending.drain_like_cpp();
-        assert_eq!(commands.len(), 5);
-        assert!(matches!(
-            commands[0],
-            SessionCommand::CreatureAttackStartLikeCpp(_)
-        ));
-        assert!(matches!(
-            commands[1],
-            SessionCommand::CreatureAttackStopLikeCpp(_)
-        ));
-        assert!(matches!(
-            commands[2],
-            SessionCommand::ReconcilePvpCombatExpiryLikeCpp(_)
-        ));
-        let SessionCommand::ApplyCreatureMeleeDamageLikeCpp(first_melee) = &commands[3] else {
-            panic!("expected first melee event");
-        };
-        let SessionCommand::ApplyCreatureMeleeDamageLikeCpp(second_melee) = &commands[4] else {
-            panic!("expected second melee event");
-        };
-        assert_eq!(first_melee.victim_health_after, 90);
-        assert_eq!(first_melee.victim_health_state_revision_after, 7);
-        assert_eq!(second_melee.victim_health_after, 75);
-        assert_eq!(second_melee.victim_health_state_revision_after, 8);
-        assert!(pending.drain_like_cpp().is_empty());
-    }
-
-    #[test]
-    fn durable_creature_runtime_commands_bound_stalled_session_memory_like_cpp() {
-        let victim = ObjectGuid::create_player(1, 902);
-        let attacker = ObjectGuid::create_world_object(
-            wow_core::guid::HighGuid::Creature,
-            0,
-            1,
-            571,
-            0,
-            9001,
-            903,
-        );
-        let command = CreatureAttackStartLikeCppCommand {
-            attacker_guid: attacker,
-            victim_guid: victim,
-            previous_victim_guid: None,
-            map_id: 571,
-            instance_id: 0,
-            packet_already_broadcast: false,
-        };
-        let mut pending = DurableCreatureRuntimeCommandsLikeCpp::default();
-        for _ in 0..MAX_DURABLE_CREATURE_RUNTIME_COMMANDS_LIKE_CPP {
-            assert!(pending.publish_attack_start_like_cpp(command.clone()));
-        }
-        assert!(!pending.publish_attack_start_like_cpp(command));
-        assert!(pending.take_overflowed_and_discard_like_cpp());
-        assert!(pending.drain_like_cpp().is_empty());
-        assert!(!pending.take_overflowed_and_discard_like_cpp());
-    }
 }
