+++ docs/architecture/ownership-and-boundaries.md
@@ -120,7 +120,7 @@ last-writer-wins policy.
 | Authenticated world connection | `wow-network::accept` and the connection task | socket/authentication task | `WorldSession` dispatch boundary | one connection; created after authentication and dropped on disconnect | Remains a network responsibility. #134 narrowed the listener to transport-owned configuration and authenticated connection outputs. |
 | Gameplay session construction aggregate | private `SessionResources` in `crates/world-server/src/session_resources.rs` | `world-server` bootstrap constructs it; the outer session callback captures and clones the `Arc` | `world-server::create_session` copies stores, registries, DB adapters, and runtime handles into each `WorldSession`; the aggregate never enters `wow-network` | process lifetime; no independent clock | #134 moved the aggregate to composition code; #136 moved the process into a library-backed composition root and extracted the private factory. Capability-specific dependency reduction remains in the physical and ownership lanes. |
 | Session directory | opaque `wow_world::session::directory::PlayerRegistry`; its `DashMap<ObjectGuid, compatibility entry>` storage is private to the owner module | named generation-aware registration/unregistration owns lifecycle; temporary state-publication compatibility methods remain counted | #192 runtime/fanout, #193 combat/loot and #194 quest/spell/movement consumers use owned bounded results and generation-checked delivery; remaining social/group access is assigned exactly in [`player-registry-consumer-map.tsv`](player-registry-consumer-map.tsv) | process lifetime; presence, incarnation and addressability are Session concerns, while gameplay fields are temporary mirrors | #150 installed the opaque lifecycle seam; #192-#194 closed their consumer slices; #195 burns down remaining production consumers; #196 removed compatibility storage APIs and fixtures; #138 relocated the already-opaque directory to the transitional `wow-world` owner, leaving the terminal `wow-session::directory` move open. |
-| Session mailbox and durable rails | `wow_network::player_registry::SessionCommand` plus ordinary and durable payloads, `SharedClientVisibleGuidsLikeCpp` and the durable creature-runtime/loot-money rails; #138 left them in that network module when the directory moved out | session/global runtime and gameplay producers enqueue commands; one session task consumes them | the owning session consumes FIFO commands and publishes acknowledgements | connection/session lifetime; queue identity, capacity, FIFO, incarnation fences, acknowledgements, and shutdown drain are observable | #191 and #190 completed their bounded protocol/rail relocations. #189 removes durable loot persistence coordination; #138 closes broad directory access; #140 owns the remaining pump and mailbox boundary without changing queue semantics. |
+| Session mailbox and durable rails | private `wow_world::session::mailbox::{protocol, durable, pump}`: `SessionCommand` plus every ordinary payload, `SharedClientVisibleGuidsLikeCpp`, the durable creature-runtime and loot-money rails, and the single pump | session/global runtime and gameplay producers enqueue commands; one session task consumes them | the owning session consumes FIFO commands and publishes acknowledgements | connection/session lifetime; queue identity, capacity, FIFO, incarnation fences, acknowledgements, and shutdown drain are observable | #191 and #190 completed their bounded protocol/rail relocations, #138 closed broad directory access, and #140 moved the complete vertical without changing queue identity, FIFO, durability, capacity, incarnation fences, acknowledgements or shutdown drain. `wow-network` now owns only transport primitives. #189 still removes durable loot persistence coordination. |
 | Group registry and pending invites | `wow_social::group::{GroupRegistry, PendingInvites}`, backed by concurrent maps inside private `model`/`invites`/`membership`/`settings`/`ready_check`/`outcome` submodules | group handlers and group timer/ready-check paths | group handlers, connected-player fanout, world-server ready-check loop | process lifetime; timed group work is driven outside the network listener | #151 created opaque facades; #197 and #198 centralized atomic transitions; #199 separated persistence/publication and closed storage; #195 adapted session addressing; #137 moved the owner into `wow-social`. The transitional `wow-network → wow-social` edge exists only because the Session mailbox still names `GroupDifficultyKindLikeCpp`; #140 removes it. |
 | Legacy creature runtime | shared `wow_world::MapManager` behind `Arc<RwLock<_>>` | production `GlobalLegacy` runtime tick plus explicit spawn/respawn bridges; `Session` writer exists only for tests and the diagnostic config override | world handlers, global runtime bridge, visibility/fanout routing | process lifetime; production startup defaults `RuntimeTickOwner` to `GlobalLegacy` and uses the configured map-update interval; session ticks read the shared owner and must skip to prevent double resolution | #188 freezes clocks, phases, and bridge behavior. #28 may perform one bounded authority cut; later cuts retire legacy behavior method-by-method under `docs/migration/adr-runtime-tick-ownership.md`. |
 | Canonical map runtime | `wow_map::MapManager` | canonical global map loop, grid/spawn/respawn paths, explicit selected legacy-result adapters | world-server orchestration and session map/player bridges | process lifetime; canonical loop uses the configured map interval; preserves the C++ `Map::Update` phase order represented by the ADR | #188 records the current phase trace. It becomes the sole map/entity authority only after each later method cut has parity tests and removes the corresponding legacy writer. |
@@ -326,6 +326,17 @@ Issue #139 retires the former 18,637-line `handlers/misc.rs` hotspot instead of 
 logical monolith: its registrations, behavior, and tests now live in private capability modules
 under `handlers/misc/`, while `misc/mod.rs` is a 164-line compatibility facade for shared packet
 types, constants, and narrow helpers. Every production capability file is below 700 lines.
+
+Issue #140 completes the same campaign for the Session mailbox and, unlike the two moves before
+it, is a net dependency reduction rather than a relocation: `wow-network` loses `wow-data`,
+`wow-loot` and `wow-social` outright, so the workspace falls from 109 to 106 edges and the baseline
+exceptions from 20 to 17 — including the transitional `wow-network -> wow-social` edge #137 had
+just declared. Raw networking now depends only on transport crates. The 1,349 production and 390
+test lines the Session logical ceiling gains are the mailbox arriving as `session/mailbox/`; the
+`WorldSession` surface itself is unchanged at 738 fields and 3,341 associated items, with no
+method added or removed by the move. The one real surface change is that 29 command handlers in
+`handlers/loot.rs` become `pub(crate)`: the pump that calls them now lives in a sibling module, and
+Rust has no narrower visibility that reaches it. They stay crate-internal and unexported.
 
 Issue #137 gives the relocated Group owner its own logical ceiling instead of letting it land
 untracked: `crates/wow-social/src/group/mod.rs` enters the table at 2,783 production and 2,511 test
