+++ tools/architecture/handler-contract-check/src/session_ownership.rs
@@ -62,6 +62,9 @@ const SESSION_RESOURCES_NAME: &str = "SessionResources";
 const SESSION_FACTORY_MODULE: &str = "crate::session_factory";
 const SESSION_FACTORY_NAME: &str = "create_session";
 const PRIVATE_WORLD_SESSION_OWNER_ROOTS: &[&str] = &["crate::handlers::misc", WORLD_SESSION_MODULE];
+/// Issue #140 relocated the Session mailbox here, so `SessionCommand` and its
+/// payload closure now live in `wow-world` rather than `wow-network`.
+const WORLD_SESSION_MAILBOX_MODULE: &str = "crate::session::mailbox";
 /// Issue #137 relocated the atomic Group owner here. `SessionCommand` still
 /// names `GroupDifficultyKindLikeCpp` in one payload, so the contract scan must
 /// reach this module or that payload type would silently leave the inventory.
@@ -1323,7 +1326,10 @@ fn collect_items(
         }
 
         if role == PackageRole::Network
-            || (role == PackageRole::World && module == WORLD_SESSION_DIRECTORY_MODULE)
+            || (role == PackageRole::World
+                && (module == WORLD_SESSION_DIRECTORY_MODULE
+                    || module == WORLD_SESSION_MAILBOX_MODULE
+                    || module.starts_with(&format!("{WORLD_SESSION_MAILBOX_MODULE}::"))))
             || (role == PackageRole::Social
                 && (module == SOCIAL_GROUP_MODULE
                     || module.starts_with(&format!("{SOCIAL_GROUP_MODULE}::"))))
