+++ tools/architecture/handler-contract-check/src/tests.rs
@@ -33,7 +33,7 @@ fn repository_handler_contract_passes() {
         .unwrap_or_else(|error| panic!("invalid repository handler contract:\n{error}"));
     assert!(report.starts_with("handler contract: PASS"), "{report}");
     assert!(report.contains("0 exact drift exceptions"), "{report}");
-    assert!(report.contains("16 #[path] modules verified"), "{report}");
+    assert!(report.contains("17 #[path] modules verified"), "{report}");
 }
 
 #[test]
